chore: update y in TrainingData from float64 to int64

+++ src/text_classifier/schema.py
@@ -10,8 +10,8 @@ from sklearn.preprocessing import LabelEncoder
 class TrainingData:
     X_train: pd.DataFrame
     X_test: pd.DataFrame
-    y_train: np.typing.NDArray[np.float64]
-    y_test: np.typing.NDArray[np.float64]
+    y_train: np.typing.NDArray[np.int64]
+    y_test: np.typing.NDArray[np.int64]
 
 
 @dataclass
